Corenet API: Move more logic away from commands
License: MIT
